Fix local mode, refactor image loading
Change-type: minor
