+++ lib/backends/fs.ts
@@ -156,16 +156,17 @@ export default class FileSystemBackend {
 	 *   console.log(contents)
 	 * })
 	 */
-	async readFile(file: string, _options?: { base64: boolean }): Promise<any> {
+	async readFile(file: string, options?: { base64: boolean }): Promise<any> {
 		try {
 			const fileContent = await fs.promises.readFile(
 				join(this.repository, file),
 				{
-					encoding: imageFileExtensions.includes(
-						file.split('.').reverse()[0]?.toLowerCase(),
-					)
-						? 'base64'
-						: 'utf8',
+					encoding:
+						imageFileExtensions.includes(
+							file.split('.').reverse()[0]?.toLowerCase(),
+						) || options?.base64
+							? 'base64'
+							: 'utf8',
 				},
 			);
 			return fileContent;
