+++ lib/plugins/logo.ts
@@ -14,158 +14,15 @@
  * limitations under the License.
  */
 
-import { promisifyAll, props } from 'bluebird';
-import { isString } from 'lodash';
-
-import { convertRemoteImageToBase64, mimeTypes } from '../utils/image';
-import { Magic as _Magic, MAGIC_MIME_TYPE } from 'mmmagic';
-const Magic = _Magic;
-const magic = promisifyAll(new Magic(MAGIC_MIME_TYPE));
-
-import { recognize } from 'tesseract.js';
-import sharp from 'sharp';
+import { props } from 'bluebird';
 import { Backend } from '../../typings/types';
+import { getLogoFromUrl } from '../utils/image';
 import {
 	convertHtmlToMD,
 	getFirstImageIndex,
 	isParagraph,
 	markdownAST,
 } from '../utils/markdown';
-
-const absoluteUrlRe = new RegExp('^(?:[a-z]+:)?//', 'i');
-
-/**
- * @summary Detects is the url is absolute or not
- * @function
- * @private
- *
- * @param {String} url - the URL
- * @returns {Boolean}
- *
- * @example
- * console.log(isAbsoluteUrl("https://google.com")) // true
- *
- * @example
- * console.log(isAbsoluteUrl("./some-image.png")) // false
- */
-const isAbsoluteUrl = (url: string): boolean => {
-	return absoluteUrlRe.test(url);
-};
-
-/**
- * @summary Detects Text from Image using Tesseract OCR
- * @function
- * @private
- *
- * @param {String} imageUrl - image url
- * @returns {Promise}
- *
- * @example
- * detectTextFromImage("https://unsplash.com/nature/example.png").then((imageText) => {
- *    console.log(imageText)
- * })
- */
-const detectTextFromImage = async (
-	imageUrl: string | Buffer,
-): Promise<string | null> => {
-	const {
-		data: { text },
-	} = await recognize(imageUrl, 'eng');
-	return text && text.length >= 3 ? text : null;
-};
-
-/**
- * @summary extracts mimeType and data from base64 encoded strings
- * @function
- * @private
- *
- * @param {String} encoded - base64 encoded string
- * @returns {{mimeType: String, data: String}}
- *
- * @example
- * const {mimeType, data} = base64MimeType(string)
- */
-const base64MimeType = (
-	encoded: string,
-): { mimeType: string | null; data: string } => {
-	if (!isString(encoded)) {
-		return { mimeType: null, data: encoded };
-	}
-
-	const regexMatch = encoded.match(
-		/data:([a-zA-Z0-9]+\/[a-zA-Z0-9-.+]+).*,(.*)/,
-	);
-
-	if (!regexMatch) {
-		return { mimeType: null, data: encoded };
-	}
-
-	return { mimeType: regexMatch[1], data: regexMatch[2] };
-};
-/* eslint-enable */
-
-/**
- * @summary Get logo Text and base64 image.
- * @function
- * @private
- *
- * @param {String} imageUrl - image link, either relative or absolute.
- * @param {Object} backend -
- * @returns {Promise}
- *
- * @example
- * getLogoFromUrl("https://unsplash.com/nature/tree.jpg").then(({logo}) => {
- *    console.log(logo);
- * })
- */
-const getLogoFromUrl = async (
-	imageUrl: string,
-	backend: Backend,
-): Promise<{ logo: { base64: string; textContent: string | null } }> => {
-	let logoText = null;
-	let base64Image = null;
-
-	if (isAbsoluteUrl(imageUrl)) {
-		base64Image = await convertRemoteImageToBase64(imageUrl);
-		const { mimeType, data } = base64MimeType(base64Image);
-		let buffer = Buffer.from(data, 'base64');
-		if (mimeType === 'image/svg' || mimeType === 'image/svg+xml') {
-			buffer = await sharp(buffer).png().toBuffer();
-			base64Image = `data:image/png;base64,${buffer.toString('base64')}`;
-		}
-		logoText = await detectTextFromImage(buffer);
-	} else {
-		let localImageUrl: string = imageUrl;
-		if (imageUrl.startsWith('./')) {
-			localImageUrl = imageUrl.replace('./', '');
-		}
-
-		// The image is local to the repo so we can fetch it via the backend
-		const files = await props({
-			logo: backend.readFile(localImageUrl),
-		});
-		let buffer = Buffer.from(files.logo, 'base64');
-
-		const mimeType = localImageUrl.split('.').reverse()[0]
-			? mimeTypes[
-					localImageUrl.split('.').reverse()[0] as keyof typeof mimeTypes
-			  ]
-			: await magic.detectAsync(buffer);
-		base64Image = `data:${mimeType};base64,${files.logo}`;
-		if (mimeType === 'image/svg' || mimeType === 'image/svg+xml') {
-			buffer = await sharp(buffer).png().toBuffer();
-			base64Image = `data:image/png;base64,${buffer.toString('base64')}`;
-		}
-		logoText = await detectTextFromImage(buffer);
-	}
-
-	return {
-		logo: {
-			base64: base64Image,
-			textContent: logoText,
-		},
-	};
-};
 
 export default async (backend: Backend) => {
 	const files = await props({
@@ -189,7 +46,7 @@ export default async (backend: Backend) => {
 	}
 
 	if (imageUrl) {
-		const { logo } = await getLogoFromUrl(imageUrl, backend);
+		const logo = await getLogoFromUrl(imageUrl, backend);
 		return { logo };
 	}
